update file list fail

+++ src/main/java/common/utility/FileHelpers.java
@@ -10,6 +10,7 @@ import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
+import java.util.List;
 
 public class FileHelpers {
     public static String getRootFolder(){
@@ -55,24 +56,26 @@ public class FileHelpers {
             f.mkdirs();
         }
     }
-    public static void deleteAllFileInFolder(String folderPath){
+    public static void deleteAllFileInFolder(List<String> reports,String folderPath){
         try {
-            File folder = new File(folderPath);
-            if (folder.isDirectory()) {
-                File[] files = folder.listFiles();
-                if (files != null) {
-                    for (File file : files) {
-                        if (!file.isDirectory()) { // Make sure it's not a directory
-                            boolean deleted = file.delete();
-                            if (deleted) {
-                                Log.info("Deleted file: " + file.getAbsolutePath());
-                            } else {
-                                Log.info("Failed to delete file: " + file.getAbsolutePath());
+            System.out.println("111111111 "+folderPath);
+            System.out.println(reports);
+                File folder = new File(folderPath);
+                if (folder.isDirectory()) {
+                    File[] files = folder.listFiles();
+                    if (files != null) {
+                        for (File file : files) {
+                            if (!file.isDirectory() && !reports.contains(file.getAbsolutePath())) { // Make sure it's not a directory
+                                boolean deleted = file.delete();
+                                if (deleted) {
+                                    Log.info("Deleted file: " + file.getAbsolutePath());
+                                } else {
+                                    Log.info("Failed to delete file: " + file.getAbsolutePath());
+                                }
                             }
                         }
                     }
                 }
-            }
         }catch (Exception e)
         {
             Log.error("The specified path is not a directory.");
